Refactor: Update recent watched data source to use Flow
This commit refactors the `RecentWatchedDataSource` and its implementations (`RecentWatchedMoviesDataSource` and `RecentWatchedTvShowsDataSource`) to return `Flow<List<T>>` instead of `List<T>` for the `getAll` method.

The `RecentWatchedRepositoryImpl` has also been updated to reflect this change, now mapping the Flow of local models to a Flow of domain entities.

Additionally, a typo in the package name `whatched` has been corrected to `watched` in `LocalDataSourceModule.kt`.
