+++ data/src/main/java/com/london/data/local/source/recent/watched/RecentWatchedDataSource.kt
@@ -1,6 +1,8 @@
 package com.london.data.local.source.recent.watched
 
+import kotlinx.coroutines.flow.Flow
+
 interface RecentWatchedDataSource<T> {
-    suspend fun getAll(): List<T>
+    suspend fun getAll(): Flow<List<T>>
     suspend fun insert(item: T)
 }
